fix: satisfy clippy --all-features across provenance and runner
Collapse host-ingress context_id validation into a match guard, drop
redundant usize casts, map ProvenanceOpsRow for drift tests, and handle
Planning in clickup transcript matching.

+++ crates/baml-agent-runner/src/services/conversation_history.rs
@@ -43,7 +43,7 @@ impl ConversationHistoryServiceImpl {
         };
         TranscriptPageRequest {
             scope,
-            limit: request.page.limit() as usize,
+            limit: request.page.limit(),
             profile,
         }
     }
